refactor(design-playwright-analyzer): extract domain modules for maintainability
Refactored design-playwright-analyzer.js (893 LOC) into focused domain modules:

- config.js (55 LOC): Breakpoints, WCAG criteria, constants
- analyzers.js (570 LOC): Core analysis methods (processFlow, consistency, accessibility, responsive, performance, interactions)
- recommendations.js (118 LOC): Recommendation generation logic
- report-generators.js (183 LOC): HTML, Markdown, JSON report output
- index.js (141 LOC): Main orchestrator with re-exports

Original file reduced to thin wrapper (56 LOC) maintaining backward compatibility.
Part of SD-LEO-REFACTOR-LARGE-FILES-003 orchestrator.
